+++ ChangeLog.md
@@ -1,2 +1,12 @@
+# 0.2.2.0
+- Add renderPCFTextColor function.
+# 0.2.1.1
+- Improve documentation in Graphics.Text.PCF.
+# 0.2.1.0
+- Add .pcf.gz support.
+# 0.2.0.0
+- Add ASCII string rendering.
+- Add PCFText type.
+- Change renderPCFText to produce a value of PCFText.
 # 0.1.0.0
-Complete font rendering and parsing support.
+- Complete font rendering and parsing support.